+++ src/plonemeeting/portal/core/browser/docgen.py
@@ -3,7 +3,7 @@ import re
 from AccessControl import Unauthorized
 from collective.behavior.talcondition.utils import WRONG_TAL_CONDITION
 from collective.documentgenerator.browser.generation_view import DocumentGenerationView
-from collective.documentgenerator.helper import DocumentGenerationHelperView
+from collective.documentgenerator.helper import DXDocumentGenerationHelperView
 from collective.timestamp.behaviors.timestamp import ITimestampableDocument
 from html import escape
 from imio.helpers.barcode import generate_barcode
@@ -19,7 +19,7 @@ from Products.CMFCore.Expression import Expression
 import copy
 
 
-class PMDocumentGenerationHelperView(DocumentGenerationHelperView):
+class PMDocumentGenerationHelperView(DXDocumentGenerationHelperView):
     """Helper view for document generation."""
 
     def __init__(self, context, request):